Rename `HiGlassComponent.js` `HiGlassComponent.jsx`
